+++ config/statamic/forms.php
@@ -48,6 +48,25 @@ return [
 
     /*
     |--------------------------------------------------------------------------
+    | Exporters
+    |--------------------------------------------------------------------------
+    |
+    | Here you may define all the available form submission exporters.
+    | You may customize the options within each exporter's array.
+    |
+    */
+
+    'exporters' => [
+        'csv' => [
+            'class' => Statamic\Forms\Exporters\CsvExporter::class,
+        ],
+        'json' => [
+            'class' => Statamic\Forms\Exporters\JsonExporter::class,
+        ],
+    ],
+
+    /*
+    |--------------------------------------------------------------------------
     | CSV Export Delimiter
     |--------------------------------------------------------------------------
     |
@@ -56,5 +75,19 @@ return [
     */
 
     'csv_delimiter' => ',',
+
+    /*
+    |--------------------------------------------------------------------------
+    | CSV Export Headings
+    |--------------------------------------------------------------------------
+    |
+    | The values to be used in the csv export header rows.
+    | Can be the field handle or the field display text.
+    |
+    | Supported values: "handle", "display"
+    |
+    */
+
+    'csv_headers' => 'display',
 
 ];